+++ internal/e2eTests/upload/e2e_from_folder_test.go
@@ -18,94 +18,6 @@ import (
 	"golang.org/x/exp/rand"
 )
 
-func TestResetImmich(t *testing.T) {
-	e2e.InitMyEnv()
-	e2e.ResetImmich(t)
-}
-
-func TestUploadFromGooglePhotos(t *testing.T) {
-	e2e.InitMyEnv()
-	e2e.ResetImmich(t)
-
-	ctx := context.Background()
-
-	c, a := cmd.RootImmichGoCommand(ctx)
-	c.SetArgs([]string{
-		"upload", "from-google-photos",
-		"--server=" + e2e.MyEnv("IMMICHGO_SERVER"),
-		"--api-key=" + e2e.MyEnv("IMMICHGO_APIKEY"),
-		// "--no-ui",
-		e2e.MyEnv("IMMICHGO_TESTFILES") + "/demo takeout/Takeout",
-	})
-
-	// let's start
-	err := c.ExecuteContext(ctx)
-	if err != nil && a.Log().GetSLog() != nil {
-		a.Log().Error(err.Error())
-	}
-}
-
-func TestUploadFromGooglePhotosZipped(t *testing.T) {
-	e2e.InitMyEnv()
-	e2e.ResetImmich(t)
-
-	ctx := context.Background()
-
-	c, a := cmd.RootImmichGoCommand(ctx)
-	c.SetArgs([]string{
-		"upload", "from-google-photos",
-		"--server=" + e2e.MyEnv("IMMICHGO_SERVER"),
-		"--api-key=" + e2e.MyEnv("IMMICHGO_APIKEY"),
-		// "--no-ui",
-		e2e.MyEnv("IMMICHGO_TESTFILES") + "/demo takeout/Takeout.zip",
-	})
-
-	// let's start
-	err := c.ExecuteContext(ctx)
-	if err != nil && a.Log().GetSLog() != nil {
-		a.Log().Error(err.Error())
-	}
-}
-
-func TestUploadFromGPInCurrent(t *testing.T) {
-	curDir, err := os.Getwd()
-	if err != nil {
-		t.Fatal(err)
-		return
-	}
-
-	t.Cleanup(func() {
-		_ = os.Chdir(curDir)
-	})
-
-	e2e.InitMyEnv()
-	e2e.ResetImmich(t)
-
-	ctx := context.Background()
-	d := e2e.MyEnv("IMMICHGO_TESTFILES") + "/demo takeout/Takeout"
-	err = os.Chdir(d)
-	if err != nil {
-		t.Fatal(err)
-		return
-	}
-
-	c, a := cmd.RootImmichGoCommand(ctx)
-	c.SetArgs([]string{
-		"upload", "from-google-photos",
-		"--server=" + e2e.MyEnv("IMMICHGO_SERVER"),
-		"--api-key=" + e2e.MyEnv("IMMICHGO_APIKEY"),
-		"--from-album-name=Duplicated album",
-		// "--no-ui",
-		".",
-	})
-
-	// let's start
-	err = c.ExecuteContext(ctx)
-	if err != nil && a.Log().GetSLog() != nil {
-		a.Log().Error(err.Error())
-	}
-}
-
 // TestUploadBurstInAlbums show the immich problem when a stack is included into an album
 // immich version v1.123.0
 func TestUploadBurstInAlbums(t *testing.T) {
